Created an abstract base class for controller tests.
