+++ web/src/test/java/com/epam/rft/atsy/web/controllers/WelcomeControllerTest.java
@@ -5,28 +5,16 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
 
-import org.junit.Before;
 import org.junit.Test;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.servlet.ViewResolver;
-import org.springframework.web.servlet.view.InternalResourceViewResolver;
 
-public class WelcomeControllerTest {
-  private static final String VIEW_PREFIX = "/WEB-INF/pages/";
-  private static final String VIEW_SUFFIX = ".jsp";
-
+public class WelcomeControllerTest extends AbstractControllerTest {
   private static final String VIEW_NAME = "candidates";
 
   private static final String REQUEST_URL = "/secure/welcome";
 
-  private MockMvc mockMvc;
-
-  @Before
-  public void setUp() {
-    mockMvc =
-        MockMvcBuilders.standaloneSetup(new WelcomeController())
-            .setViewResolvers(viewResolver()).build();
+  @Override
+  protected Object[] controllersUnderTest() {
+    return new Object[] { new WelcomeController() };
   }
 
   @Test
@@ -35,15 +23,5 @@ public class WelcomeControllerTest {
         .andExpect(status().isOk())
         .andExpect(view().name(VIEW_NAME))
         .andExpect(forwardedUrl(VIEW_PREFIX + VIEW_NAME + VIEW_SUFFIX));
-  }
-
-
-  private ViewResolver viewResolver() {
-    InternalResourceViewResolver viewResolver = new InternalResourceViewResolver();
-
-    viewResolver.setPrefix(VIEW_PREFIX);
-    viewResolver.setSuffix(VIEW_SUFFIX);
-
-    return viewResolver;
   }
 }
